+++ modernweb/packages/cli/src/interpreter/command.js
@@ -1,4 +1,23 @@
+const Context = require("./context");
+
 class Command {
+  /**
+   * @type {Context?}
+   * @protected
+   */
+  context = null;
+  /**
+   * @type {String?}
+   * @protected
+   */
+  root = null;
+
+  /** @param {Context} context */
+  constructor(context) {
+    this.context = context;
+    this.root = process.cwd();
+  }
+
   execute() {
     throw new Error("Not Implemented!");
   }
